+++ Kernel/System/WebUserAgent.pm
@@ -76,6 +76,7 @@ Simple GET request:
     my %Response = $WebUserAgentObject->Request(
         URL => 'http://example.com/somedata.xml',
         SkipSSLVerification => 1, # (optional)
+        NoLog               => 1, # (optional)
     );
 
 Or a POST request; attributes can be a hashref like this:
@@ -85,6 +86,7 @@ Or a POST request; attributes can be a hashref like this:
         Type => 'POST',
         Data => { Attribute1 => 'Value', Attribute2 => 'Value2' },
         SkipSSLVerification => 1, # (optional)
+        NoLog               => 1, # (optional)
     );
 
 alternatively, you can use an arrayref like this:
@@ -94,6 +96,7 @@ alternatively, you can use an arrayref like this:
         Type => 'POST',
         Data => [ Attribute => 'Value', Attribute => 'OtherValue' ],
         SkipSSLVerification => 1, # (optional)
+        NoLog               => 1, # (optional)
     );
 
 returns
@@ -114,6 +117,7 @@ You can even pass some headers
             Content_Type  => 'text/json',
         },
         SkipSSLVerification => 1, # (optional)
+        NoLog               => 1, # (optional)
     );
 
 If you need to set credentials
@@ -129,6 +133,7 @@ If you need to set credentials
             Location => 'ftp.otrs.org:80',
         },
         SkipSSLVerification => 1, # (optional)
+        NoLog               => 1, # (optional)
     );
 
 =cut
@@ -215,10 +220,14 @@ sub Request {
     }
 
     if ( !$Response->is_success() ) {
-        $Kernel::OM->Get('Kernel::System::Log')->Log(
-            Priority => 'error',
-            Message  => "Can't perform $Param{Type} on $Param{URL}: " . $Response->status_line(),
-        );
+
+        if (!$Param{NoLog}) {
+            $Kernel::OM->Get('Kernel::System::Log')->Log(
+                Priority => 'error',
+                Message  => "Can't perform $Param{Type} on $Param{URL}: " . $Response->status_line(),
+            );
+        }
+
         return (
             Status => $Response->status_line(),
         );